requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
openai>=1.0.0
PyYAML>=6.0.1